Silence a warning.
git-svn-id: http://root.cern.ch/svn/root/trunk@49312 27541ba8-7e3a-0410-8455-c3a389f83636

+++ lib/Interpreter/IncrementalParser.cpp
@@ -217,7 +217,7 @@ namespace cling {
       if (T->isNestedTransaction()) {
         // We need to remove the marker from its parent.
         Transaction* ParentT = T->getParent();
-        for (int i = 0; i < ParentT->size(); ++i)
+        for (size_t i = 0; i < ParentT->size(); ++i)
           if ((*ParentT)[i].m_DGR.isNull())
             ParentT->erase(i);
       }
